# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FinAegis Core Banking API",
        "description": "Open Source Core Banking as a Service - A modern, scalable, and secure core banking platform built with Laravel 12, featuring event sourcing, domain-driven design, workflow orchestration, and quantum-resistant security measures.",
        "contact": {
            "name": "FinAegis Support",
            "email": "support@finaegis.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/assets": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "List all supported assets",
                "description": "Get a list of all assets supported by the platform, including fiat currencies, cryptocurrencies, and commodities",
                "operationId": "listAssets",
                "parameters": [
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Filter by active status",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by asset type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "fiat",
                                "crypto",
                                "commodity"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by code or name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "code": {
                                                        "type": "string",
                                                        "example": "USD"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "US Dollar"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "enum": [
                                                            "fiat",
                                                            "crypto",
                                                            "commodity"
                                                        ]
                                                    },
                                                    "symbol": {
                                                        "type": "string",
                                                        "example": "$"
                                                    },
                                                    "precision": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "is_active": {
                                                        "type": "boolean"
                                                    },
                                                    "metadata": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "active": {
                                                    "type": "integer"
                                                },
                                                "types": {
                                                    "properties": {
                                                        "fiat": {
                                                            "type": "integer"
                                                        },
                                                        "crypto": {
                                                            "type": "integer"
                                                        },
                                                        "commodity": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "scheme": "bearer"
            }
        }
    },
...
}